move BlogCategoryMenu to a separate menu.py
